code refactoring2
