+++ web/src/main/java/com/ifive/front/domain/Past/MusicalPastDTO.java
@@ -1,5 +1,4 @@
-package com.ifive.front.Present;
-
+package com.ifive.front.domain.Past;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
@@ -13,7 +12,7 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 @JsonIgnoreProperties(ignoreUnknown = true) // DTO에는 numb가없어서 무시하게함
-public class MusicalPresentDTO {
+public class MusicalPastDTO {
 
     private Integer musicalId;
     private String title;
