+++ web/src/main/java/com/ifive/front/domain/Past/MusicalPastService.java
@@ -1,4 +1,4 @@
-package com.ifive.front.Past;
+package com.ifive.front.domain.Past;
 
 import java.util.List;
