+++ web/src/main/java/com/ifive/front/domain/Present/MusicalPresent.java
@@ -1,4 +1,4 @@
-package com.ifive.front.Present;
+package com.ifive.front.domain.Present;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
