+++ web/src/main/java/com/ifive/front/domain/Present/MusicalPresentController.java
@@ -1,4 +1,4 @@
-package com.ifive.front.Present;
+package com.ifive.front.domain.Present;
 
 import java.util.HashMap;
 import java.util.List;
@@ -9,7 +9,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import lombok.extern.slf4j.Slf4j;
