+++ web/src/main/resources/templates/fragment/common_tag2.html
@@ -1,0 +1,48 @@
+<!DOCTYPE html>
+<body>
+    <script>
+        function redirectToUrlWithTag(tag) {
+            const buttonContainer = document.querySelector('.sub_container');
+
+            const tags = new Map([
+                ['love', '사랑'],
+                ['opera', '오페라'],
+                ['child', '가족'],
+                ['friend', '친구'],
+                ['art', '예술'],
+                ['magic', '판타지'],
+                ['person', '휴머니즘'],
+                ['history', '역사'],
+                ['fear', '공포'],
+                ['laugh', '코미디'],
+                ['sad', '슬픔']
+            ]);
+
+            const createButton = (tag, condition) => {
+                if (condition) {
+                    const button = document.createElement('button');
+                    button.type = 'button';
+                    button.className = 'btn btn-primary btn-lg';
+                    button.onclick = () => redirectToUrlWithTag(tag);
+                    button.textContent = tags.get(tag);
+                    buttonContainer.appendChild(button);
+                }
+            };
+
+            // 태그가 주어진 경우, 해당 태그와 조건에 따라 버튼 생성
+            if (tag) {
+                createButton(tag, true);
+            }
+            // 태그가 주어지지 않은 경우, 모든 태그에 대한 버튼 생성
+            else {
+                tags.forEach((value, key) => {
+                    createButton(key, true);
+                });
+            }
+        }
+
+        // 초기 로드 시 태그가 주어지지 않은 상태로 호출
+        redirectToUrlWithTag();
+    </script>
+</body>
+</html>
